% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-args.R
\name{check_method_dots}
\alias{check_method_dots}
\title{Check that \code{...} passed to a method is empty}
\usage{
check_method_dots(...)
}
\arguments{
\item{...}{\code{...} passed on in an S3 method.}
}
\description{
It's good practice to use \code{...} in S3 generic so that the methods can
take additional arguments. However, since \code{...} silently swallow mis-nammed
arguments, it's friendly to tell the user when there's something in \code{...}
that you didn't expect
}
\examples{
foo <- function(x, ...) UseMethod("foo")

foo.character <- function(x, ..., truncate = 10) {
  substr(x, 1, truncate)
}
# Silently returns wrong result because misspelled argument name is
# swallowed by ...
foo("text", trncate = 5)

foo.character <- function(x, ..., truncate = 10) {
  check_method_dots(...)
  substr(x, 1, truncate)
}
# Throws an error because `...` is not empty
\dontrun{
foo("text", trncate = 5)
}
}
