% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date.R
\name{Date}
\alias{Date}
\alias{new_Date}
\title{Date}
\usage{
Date(year, month, day)

new_Date(x)
}
\arguments{
\item{year, month, day}{Numeric vectors defining the date.}

\item{x}{A numeric vector}
}
\value{
A S3 object with class Date
}
\description{
A \code{Date} is a double vector. The value represent the number of days
since the Unix "epoch", 1970-01-01. It has no attributes.
}
\note{
The \code{Date()} helper is rather inefficient as it uses \code{\link[=ISOdate]{ISOdate()}}
which works by pasting together a string and then passing to \code{\link[=strptime]{strptime()}}.
A more efficient implementation is available \code{\link[lubridate:make_date]{lubridate::make_date()}}
}
\examples{
new_Date(as.double(1:10))
}
