% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_class.R
\name{implicit_class}
\alias{implicit_class}
\title{Compute the implicit S3 class of a base type}
\usage{
implicit_class(x)
}
\arguments{
\item{x}{A primitive type}
}
\description{
Compute the implicit S3 class of a base type
}
\examples{
implicit_class(NULL)

implicit_class(logical())
implicit_class(integer())
implicit_class(numeric())
implicit_class(character())

implicit_class(matrix())
implicit_class(matrix(1))

implicit_class(array())
implicit_class(array(1))
}
