% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{new_data.frame}
\alias{new_data.frame}
\title{data.frame}
\usage{
new_data.frame(x, row.names = NULL)
}
\arguments{
\item{x}{A list.}

\item{row.names}{Can be a integer or character vector the same length as
the data frame. Alternatively can be a numeric vector of length 2,
as created by \code{\link[=.set_row_names]{.set_row_names()}}.}
}
\description{
A \code{data.frame} is a list with "row.names" attribute. Each element of the
list must be named, and of the same length.
}
