% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date.R
\name{new_Date}
\alias{new_Date}
\title{Date}
\usage{
new_Date(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A S3 object with class Date
}
\description{
A \code{Date} is a double vector. The value represent the number of days
since the Unix "epoch", 1970-01-01. It has no attributes
}
\examples{
new_Date(as.double(1:10))
}
