% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSIXct.R
\name{new_POSIXct}
\alias{new_POSIXct}
\title{POSIXct}
\usage{
new_POSIXct(x, tzone = "")
}
\arguments{
\item{x}{A double vector}

\item{tzone}{The time zone, a character vector of length one.
\code{""} is a dummy value that means to use the local time zone.
\code{\link[=OlsonNames]{OlsonNames()}} lists all other valid names}
}
\description{
A \code{POSIXct} is a double vector. Its value represents the number of
seconds since the Unix "Epoch", 1970-01-01. It has one attribute:
the timezone.
}
\examples{
new_POSIXct(as.double(1:10))
}
