% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_s3.R
\name{new_s3}
\alias{new_s3}
\alias{new_s3_lgl}
\alias{new_s3_int}
\alias{new_s3_dbl}
\alias{new_s3_chr}
\alias{new_s3_lst}
\title{S3 vector constructors}
\usage{
new_s3_lgl(x, ..., class)

new_s3_int(x, ..., class)

new_s3_dbl(x, ..., class)

new_s3_chr(x, ..., class)

new_s3_lst(x, ..., class)
}
\arguments{
\item{x}{A vector. An error will be throw if the type does not
match the function suffix.}

\item{...}{Name-value pairs which form the fields (attributes) of
the class}

\item{class}{Class vector}
}
\description{
These functions aid in constructing common types of S3 classes.
Since S3 does not provide a formal notion of a class, you should always
provide a constructor function that wraps one of these functions,
checking the type of all attributes.
}
\examples{
new_s3_dbl(0, class = "Date")
new_s3_dbl(0, class = "POSIXct")
}
