% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_s3.R
\name{new_s3_scalar}
\alias{new_s3_scalar}
\title{S3 scalar constructor}
\usage{
new_s3_scalar(..., class)
}
\arguments{
\item{...}{Name-value pairs which form the fields (list-elements) of
the class}

\item{class}{Class vector}
}
\description{
Used to construct "scalar" S3 objects which use a list with
named elements
}
\examples{
new_s3_scalar(a = 1, b = 2, class = "my_class")
}
