% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otype.R
\name{otype}
\alias{otype}
\title{Determine the type of an object}
\usage{
otype(x)
}
\arguments{
\item{x}{An object}
}
\description{
Tells you if you're dealing with an base, S3, S4, RC, or R6 object.
}
\examples{
otype(1:10)
otype(mtcars)
}
