% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSIXct.R
\name{POSIXct}
\alias{POSIXct}
\alias{new_POSIXct}
\title{POSIXct}
\usage{
POSIXct(year, month, day, hour, minute, sec, tzone = "")

new_POSIXct(x, tzone = "")
}
\arguments{
\item{year, month, day}{Numeric vectors defining the date.}

\item{hour, minute, sec}{Numeric vectors defining the time.}

\item{tzone}{The time zone, a character vector of length one.
\code{""} is a dummy value that means to use the local time zone.
\code{\link[=OlsonNames]{OlsonNames()}} lists all other valid names}

\item{x}{A double vector}
}
\description{
A \code{POSIXct} is a double vector. Its value represents the number of
seconds since the Unix "Epoch", 1970-01-01. It has one attribute:
the timezone.
}
\note{
The \code{POSIXct()} help is rather inefficient as it uses \code{\link[=ISOdatetime]{ISOdatetime()}}
which works by pasting together a string and then passing to \code{\link[=strptime]{strptime()}}.
A more efficient implementation is available \code{\link[lubridate:make_datetime]{lubridate::make_datetime()}}
}
\examples{
new_POSIXct(as.double(1:10))
}
