% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct an S3 class from a template}
\usage{
reconstruct(new, old)
}
\arguments{
\item{new}{Freshly created object (typically a bare vector)}

\item{old}{Existing object to use an template}
}
\description{
This is useful in methods that use \code{\link[=NextMethod]{NextMethod()}} to get the default
behaviour, but need to copy over attributes afterwards. Typically,
a reconstruct method should use the constructor rather than copying
attributes.
}
\section{S3 dispatch}{


This is an S3 generic that (unusually) dispatches on the second argument,
\code{old}.
}

