% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{s3_class}
\alias{s3_class}
\title{Compute the S3 class of an object}
\usage{
s3_class(x)
}
\arguments{
\item{x}{A primitive type}
}
\description{
Compared to \code{\link[=class]{class()}}, this always returns the class vector that is
used for dispatch. This is most important for objects where the
class attribute has not been set.
}
\examples{
s3_class(NULL)

s3_class(logical())
s3_class(integer())
s3_class(numeric())
s3_class(character())

s3_class(matrix())
s3_class(matrix(1))

s3_class(array())
s3_class(array(1))
}
