% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{s3_dispatch}
\alias{s3_dispatch}
\title{Illustrate S3 dispatch}
\usage{
s3_dispatch(call, env = parent.frame())
}
\arguments{
\item{call}{Example call to S3 method}

\item{env}{Environment in which to evaluate call}
}
\description{
Illustrate S3 dispatch
}
\examples{
x <- Sys.time()
s3_dispatch(print(x))
s3_dispatch(is.numeric(x))
s3_dispatch(as.Date(x))
s3_dispatch(sum(x))
}
