% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-method.R
\name{s3_get_method}
\alias{s3_get_method}
\title{Find S3 method from its name}
\usage{
s3_get_method(name)
}
\arguments{
\item{name}{A string or unquoted symbol}
}
\value{
A function, or an error stating why the method could not be
found
}
\description{
Find S3 method from its name
}
\examples{
s3_get_method(mean.Date)
s3_get_method(weighted.mean.Date)
}
