% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloop-package.R
\docType{package}
\name{sloop-package}
\alias{sloop}
\alias{sloop-package}
\title{sloop: Helpers for 'OOP' in R}
\description{
A collection of helper functions designed to help you to better understand object oriented programming in R, particularly using 'S3'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/sloop}
  \item \url{https://sloop.r-lib.org}
  \item Report bugs at \url{https://github.com/r-lib/sloop/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
