% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{new_factor}
\alias{new_factor}
\title{Factor}
\usage{
new_factor(x, levels)
}
\arguments{
\item{x}{An integer vector}

\item{levels}{A character vector of levels}
}
\description{
A \code{factor} is an integer with attribute levels. There should be one
level for each integer between 1 and \code{max(x)}. Note that \code{new_factor()} only
checks types; it does not assert that the values are a valid factor.
}
